Lots of changes.
Will now sort the elements in the saved variables files, so that git can easily construct deltas for updated files. I trust dump.lua well enough now not to corrupt peoples data.

Can now detect if a file is actually a QuestHelper SaveVariables file, and not copy it to the Input directory if it isn't.

Added a bash script named decompress.sh. Should go through all the zip, rar, and 7z archives it finds in the LocalInput folder, save any text/archives it finds in the LocalInput with a name based on its SHA1, and then delete the archive.

With the above changes combined, you should be able to just dump every attachment in your inbox to LocalInput, run decompress.sh, then run the build script, and then delete everything in LocalInput. On that note, there's an AddOn for Thunderbird called Attachment Extractor that can save all your attachment to some Directory, and automatically give files unique names.

In an attempt to reduce the amount of time required to check if a file obsoletes another, QuestHelper generates a random UID for each player's account, along with the last time it was used. Any updates to files we already have however are going to be appear as entirely new files, however.