Fixed menus to make sure they show up over the map window.
Seems that the fix to avoid parenting to the Cartographer map frame caused
the parent-less menus to revert to strata "MEDIUM".  I don't know that
this is the RIGHT fix, but it'll do for now.
Also added names to the menu frames so WheresMyCow wouldn't blow up when
examining them (WheresMyCow is a frame analysis tool.  Wierd name, I know).