Messed about with menus.
Made menus children of UIParent.
Set the correct strata for menuitems, so they could still be clicked.

Tried removing the parent from text strings while getting their size, I'm assuming scaling is messing up their returned dimensions.
It's still not always right, but it seems to be right more often that it was before, at least. But that might just be in my head.