ReleaseTable should be much faster now.
I think releasing tables was probably the bottleneck when recalculating the pathing information and releasing all those cached distances.

Unused tables can be garbage collected, as well as the tables used in caching distances. This will screw up the counts for the number of tables in use, but I don't care.