Some bug fixes.
Fixed a call to TravelTime that had the wrong position passed to it.

Will only advance recheck_position if the node being checked doesn't move, so that nodes don't get skipped.

My check for filtered zones only works routing for that node had already been setup, so I added 'i' for questhelper's zone indexes to
the zone node lists, and added a 'zones' table to objectives when they are prepared, which will continue to exist after routing for
them is finished, its keys are zone indexes it is known to exist in.

I added the zones table because otherwise I'd need to keep preparing and unpreparing the objective each time I wanted to test to if
it should be added to the route.

I also updated the routetest script to work with the changes made to routing.lau.