Made quest tracker movable.
I made the minimize button transparent instead of hiding it. If the game decided to start dragging after the mouse was moved off of
the tracker and it had gotten hidden, the button would get stuck to the cursor.

I added a 'treset' slash command to reset the position if needed, and added an option to reset the position in the settings menu.