Added filter /qh tracked
Added the filter for users who just want help on the quests they choose.
Kinda seemed like a one-size-fits-all filter, when I stopped to think
about it.
As part of this, moved the code for checking if an objective is watched
from tracker.lua to objective.lua, where it became a member function.